What does this PR do?
  Add the view single blog feature

Tasks to be accomplished

  - Add the blog view route
  - Add the view controller function with integrated target existence check

How to manually test this feature

  - Run the command git checkout -b ft-api-get-blog-167449792
  - Run git pill ft-api-get-blog-167449792
  - On postman, run a GET/http://localhost:3000/api/article/2 request
    Change the id value to a large value, say 100000 and see the 404 response
  - Alternatively, open your browser and type localhost:3000/api/article/2 in the address bar

Related PT story

  [finishes #167449792](https://www.pivotaltracker.com/story/show/167449792)
